Refactor restarts and Add restart countdown
